Создает модули проекта
