+++ index.html
@@ -235,7 +235,7 @@
     </template>
 
     <script src="/js/functions.js"></script>
-    <script src="/js/main.js"></script>
+    <script src="/js/main.js" type="module"></script>
 
 </body>
 </html>
