+++ js/const.js
@@ -1,0 +1,36 @@
+const MAX_PHOTOS = 25;
+const MIN_COMMENTS = 0;
+const MAX_COMMENTS = 30;
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const MAX_AVATARS = 6;
+const MIN_VALUE = 1;
+
+const NAMES = [
+  'Иван',
+  'Хуан Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
+const MESSAGES = [
+  'Всё отлично!В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию,хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+const DESCRIPTIONS = [
+  'Солнечный закат',
+  'Горячий кофе',
+  'Радужный дождь',
+  'Солнечный день',
+  'Зеленый луг',
+  'Красивый океан',
+  'Волшебный лес'
+];
+
+export{DESCRIPTIONS,MAX_AVATARS,MAX_COMMENTS,MAX_LIKES,MAX_PHOTOS,MESSAGES,MIN_COMMENTS,MIN_LIKES,MIN_VALUE,NAMES};
